chore: Makes App injectable, moves injections to composition root/main
